% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterengine.R
\name{ClusterEngine}
\alias{ClusterEngine}
\title{ClusterEngine for Managing Clustering Algorithms}
\description{
ClusterEngine for Managing Clustering Algorithms

ClusterEngine for Managing Clustering Algorithms
}
\details{
This R6 class provides a common wrapper around multiple clustering algorithms.
It stores the input dataset, the chosen method, and the resulting fitted model.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{A data frame of variables to cluster.}

\item{\code{method}}{The clustering algorithm to use: "varclus", "kmeans", or "acm_cah".}

\item{\code{n_clusters}}{Number of clusters (NULL for automatic selection if applicable).}

\item{\code{model}}{The fitted clustering model object.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ClusterEngine-new}{\code{ClusterEngine$new()}}
\item \href{#method-ClusterEngine-fit}{\code{ClusterEngine$fit()}}
\item \href{#method-ClusterEngine-predict}{\code{ClusterEngine$predict()}}
\item \href{#method-ClusterEngine-print}{\code{ClusterEngine$print()}}
\item \href{#method-ClusterEngine-summary}{\code{ClusterEngine$summary()}}
\item \href{#method-ClusterEngine-clone}{\code{ClusterEngine$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClusterEngine-new"></a>}}
\if{latex}{\out{\hypertarget{method-ClusterEngine-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new ClusterEngine
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClusterEngine$new(data, method, n_clusters = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{A data frame containing the variables to cluster}

\item{\code{method}}{Character string specifying the clustering algorithm}

\item{\code{n_clusters}}{Number of clusters (default NULL for auto)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClusterEngine-fit"></a>}}
\if{latex}{\out{\hypertarget{method-ClusterEngine-fit}{}}}
\subsection{Method \code{fit()}}{
Fit the selected clustering algorithm
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClusterEngine$fit()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The fitted model object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClusterEngine-predict"></a>}}
\if{latex}{\out{\hypertarget{method-ClusterEngine-predict}{}}}
\subsection{Method \code{predict()}}{
Predict new variable's class using the fitted model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClusterEngine$predict(new_data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{new_data}}{A data frame of new variable to predict}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Predicted cluster assignments or illustrative analysis
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClusterEngine-print"></a>}}
\if{latex}{\out{\hypertarget{method-ClusterEngine-print}{}}}
\subsection{Method \code{print()}}{
Print model brief summary
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClusterEngine$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Output of the model's print method
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClusterEngine-summary"></a>}}
\if{latex}{\out{\hypertarget{method-ClusterEngine-summary}{}}}
\subsection{Method \code{summary()}}{
Display details of the fitted clustering model results
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClusterEngine$summary()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Output of the model's summary results
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClusterEngine-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ClusterEngine-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClusterEngine$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
