% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crime}
\alias{crime}
\title{US Crime Statistics (Quantitative)}
\format{
A data frame with numeric variables
}
\source{
Data Analysis and Statistical Learning (DASL) project
}
\usage{
crime
}
\description{
Dataset containing crime statistics and demographic information for US states.
All variables are quantitative.

\strong{Tested with}: \code{KMeansVariablesQuant}
}
\examples{
\dontrun{
# Load dataset
data(crime)
str(crime)

# K-means variable clustering
km <- KMeansVariablesQuant$new(k = 3)
km$fit(crime)
km$summary()
km$plot_biplot()
}
}
\keyword{datasets}
