% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_clusters.R
\name{kmeans_elbow}
\alias{kmeans_elbow}
\title{Determine optimal number of clusters K using elbow method (K-Means Variables)}
\usage{
kmeans_elbow(X_num, k_range = 2:10, n_init = 20, seed = NULL)
}
\arguments{
\item{X_num}{Numeric data matrix or data.frame}

\item{k_range}{Range of k values to test (default: 2:10)}

\item{n_init}{Number of random initializations per k (default: 20)}

\item{seed}{Random seed for reproducibility}
}
\value{
List with optimal_k, plot function, results data.frame
}
\description{
Determine optimal number of clusters K using elbow method (K-Means Variables)
}
