% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans.R
\name{KMeansVariablesQuant}
\alias{KMeansVariablesQuant}
\title{Elbow method to determine optimal number of clusters}
\description{
Elbow method to determine optimal number of clusters

Elbow method to determine optimal number of clusters
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-KMeansVariablesQuant-new}{\code{KMeansVariablesQuant$new()}}
\item \href{#method-KMeansVariablesQuant-check_data}{\code{KMeansVariablesQuant$check_data()}}
\item \href{#method-KMeansVariablesQuant-fit}{\code{KMeansVariablesQuant$fit()}}
\item \href{#method-KMeansVariablesQuant-predict}{\code{KMeansVariablesQuant$predict()}}
\item \href{#method-KMeansVariablesQuant-summary}{\code{KMeansVariablesQuant$summary()}}
\item \href{#method-KMeansVariablesQuant-compute_elbow}{\code{KMeansVariablesQuant$compute_elbow()}}
\item \href{#method-KMeansVariablesQuant-plot_elbow}{\code{KMeansVariablesQuant$plot_elbow()}}
\item \href{#method-KMeansVariablesQuant-plot_correlation_circle}{\code{KMeansVariablesQuant$plot_correlation_circle()}}
\item \href{#method-KMeansVariablesQuant-plot_variable_map}{\code{KMeansVariablesQuant$plot_variable_map()}}
\item \href{#method-KMeansVariablesQuant-plot_center_map}{\code{KMeansVariablesQuant$plot_center_map()}}
\item \href{#method-KMeansVariablesQuant-elbow}{\code{KMeansVariablesQuant$elbow()}}
\item \href{#method-KMeansVariablesQuant-print}{\code{KMeansVariablesQuant$print()}}
\item \href{#method-KMeansVariablesQuant-clone}{\code{KMeansVariablesQuant$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KMeansVariablesQuant-new"></a>}}
\if{latex}{\out{\hypertarget{method-KMeansVariablesQuant-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KMeansVariablesQuant$new(
  k = 3,
  max_iter = 100,
  n_init = 10,
  tol = 1e-04,
  seed = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KMeansVariablesQuant-check_data"></a>}}
\if{latex}{\out{\hypertarget{method-KMeansVariablesQuant-check_data}{}}}
\subsection{Method \code{check_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KMeansVariablesQuant$check_data(X, min_vars = 2)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KMeansVariablesQuant-fit"></a>}}
\if{latex}{\out{\hypertarget{method-KMeansVariablesQuant-fit}{}}}
\subsection{Method \code{fit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KMeansVariablesQuant$fit(X, k = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KMeansVariablesQuant-predict"></a>}}
\if{latex}{\out{\hypertarget{method-KMeansVariablesQuant-predict}{}}}
\subsection{Method \code{predict()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KMeansVariablesQuant$predict(X_new)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KMeansVariablesQuant-summary"></a>}}
\if{latex}{\out{\hypertarget{method-KMeansVariablesQuant-summary}{}}}
\subsection{Method \code{summary()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KMeansVariablesQuant$summary(print_summary = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KMeansVariablesQuant-compute_elbow"></a>}}
\if{latex}{\out{\hypertarget{method-KMeansVariablesQuant-compute_elbow}{}}}
\subsection{Method \code{compute_elbow()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KMeansVariablesQuant$compute_elbow(k_range = 2:10)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KMeansVariablesQuant-plot_elbow"></a>}}
\if{latex}{\out{\hypertarget{method-KMeansVariablesQuant-plot_elbow}{}}}
\subsection{Method \code{plot_elbow()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KMeansVariablesQuant$plot_elbow(k_range = 2:10)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KMeansVariablesQuant-plot_correlation_circle"></a>}}
\if{latex}{\out{\hypertarget{method-KMeansVariablesQuant-plot_correlation_circle}{}}}
\subsection{Method \code{plot_correlation_circle()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KMeansVariablesQuant$plot_correlation_circle()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KMeansVariablesQuant-plot_variable_map"></a>}}
\if{latex}{\out{\hypertarget{method-KMeansVariablesQuant-plot_variable_map}{}}}
\subsection{Method \code{plot_variable_map()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KMeansVariablesQuant$plot_variable_map()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KMeansVariablesQuant-plot_center_map"></a>}}
\if{latex}{\out{\hypertarget{method-KMeansVariablesQuant-plot_center_map}{}}}
\subsection{Method \code{plot_center_map()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KMeansVariablesQuant$plot_center_map()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KMeansVariablesQuant-elbow"></a>}}
\if{latex}{\out{\hypertarget{method-KMeansVariablesQuant-elbow}{}}}
\subsection{Method \code{elbow()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KMeansVariablesQuant$elbow(k_range = 2:10, n_init = 20, plot = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{k_range}}{Range of k values to test (default: 2:10)}

\item{\code{n_init}}{Number of initializations per k (default: 20)}

\item{\code{plot}}{Logical, whether to display the plot (default: TRUE)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
List with optimal_k, results data.frame, and plot function
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KMeansVariablesQuant-print"></a>}}
\if{latex}{\out{\hypertarget{method-KMeansVariablesQuant-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KMeansVariablesQuant$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KMeansVariablesQuant-clone"></a>}}
\if{latex}{\out{\hypertarget{method-KMeansVariablesQuant-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KMeansVariablesQuant$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
