% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{loisirs}
\alias{loisirs}
\title{Leisure Activities Dataset (Qualitative)}
\format{
A data frame with factor variables
}
\usage{
loisirs
}
\description{
Dataset containing qualitative variables about leisure activities.
All variables are categorical.

\strong{Tested with}: \code{ClustModalities} (ACM-CAH method)
}
\examples{
\dontrun{
# Load dataset
data(loisirs)
str(loisirs)

# ACM-CAH clustering
cm <- ClustModalities$new(method = "acm", n_axes = 5)
cm$fit(loisirs, k = 3)
cm$summary()
cm$plot_factorial_map()
cm$plot_dendrogram()

# DICE method
cm_dice <- ClustModalities$new(method = "dice")
cm_dice$fit(loisirs, k = 3)
cm_dice$plot_dendrogram()
}
}
\keyword{datasets}
