% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Launch the Shiny app for exploring ClusteringVariables R Package}
\usage{
run_app()
}
\description{
This function runs the Shiny application included in the package.
Shiny is in Suggests, so it must be installed first.
}
