% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uscrime}
\alias{uscrime}
\title{US Crime Dataset (Quantitative)}
\format{
A data frame with numeric variables
}
\usage{
uscrime
}
\description{
US crime dataset with quantitative variables.

\strong{Tested with}: \code{VarClus}
}
\examples{
\dontrun{
# Load dataset
data(uscrime)
str(uscrime)

# VarClus hierarchical clustering
vc <- VarClus$new(similarity = "pearson")
vc$fit(uscrime)
vc$summary()
dend_plot <- vc$get_dendrogram()
dend_plot()
}
}
\keyword{datasets}
