% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_clusters.R
\name{varclus_elbow}
\alias{varclus_elbow}
\title{Determine optimal number of clusters K using elbow method (VarClus)}
\usage{
varclus_elbow(X_num, similarity = "spearman")
}
\arguments{
\item{X_num}{Numeric data matrix}

\item{similarity}{Similarity measure ("spearman" or "pearson")}
}
\value{
List with optimal_k and plot function
}
\description{
Determine optimal number of clusters K using elbow method (VarClus)
}
