% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_clusters.R
\name{acm_cah_elbow}
\alias{acm_cah_elbow}
\title{Determine optimal number of clusters K using elbow method (ACM-CAH)}
\usage{
acm_cah_elbow(X_quali, method = "acm", k_max = 10)
}
\arguments{
\item{X_quali}{Qualitative data (data.frame with factors)}

\item{method}{"dice" or "acm" (default: "acm")}

\item{k_max}{Maximum number of clusters to consider (default: 10)}
}
\value{
List with optimal_k, plot function, results data.frame
}
\description{
Determine optimal number of clusters K using elbow method (ACM-CAH)
}
