% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acm_cah.R
\name{ClustModalities}
\alias{ClustModalities}
\title{Elbow method with automatic detection of optimal k}
\description{
Elbow method with automatic detection of optimal k

Elbow method with automatic detection of optimal k
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ClustModalities-new}{\code{ClustModalities$new()}}
\item \href{#method-ClustModalities-check_data}{\code{ClustModalities$check_data()}}
\item \href{#method-ClustModalities-compute_dice_distances}{\code{ClustModalities$compute_dice_distances()}}
\item \href{#method-ClustModalities-fit}{\code{ClustModalities$fit()}}
\item \href{#method-ClustModalities-compute_elbow}{\code{ClustModalities$compute_elbow()}}
\item \href{#method-ClustModalities-plot_elbow}{\code{ClustModalities$plot_elbow()}}
\item \href{#method-ClustModalities-elbow}{\code{ClustModalities$elbow()}}
\item \href{#method-ClustModalities-plot_dendrogram}{\code{ClustModalities$plot_dendrogram()}}
\item \href{#method-ClustModalities-plot_factor_map}{\code{ClustModalities$plot_factor_map()}}
\item \href{#method-ClustModalities-plot_scree}{\code{ClustModalities$plot_scree()}}
\item \href{#method-ClustModalities-plot_contrib}{\code{ClustModalities$plot_contrib()}}
\item \href{#method-ClustModalities-project_numeric}{\code{ClustModalities$project_numeric()}}
\item \href{#method-ClustModalities-cluster_table}{\code{ClustModalities$cluster_table()}}
\item \href{#method-ClustModalities-predict}{\code{ClustModalities$predict()}}
\item \href{#method-ClustModalities-print}{\code{ClustModalities$print()}}
\item \href{#method-ClustModalities-summary}{\code{ClustModalities$summary()}}
\item \href{#method-ClustModalities-clone}{\code{ClustModalities$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClustModalities-new"></a>}}
\if{latex}{\out{\hypertarget{method-ClustModalities-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClustModalities$new(method = c("dice", "acm"), n_axes = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClustModalities-check_data"></a>}}
\if{latex}{\out{\hypertarget{method-ClustModalities-check_data}{}}}
\subsection{Method \code{check_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClustModalities$check_data(X)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClustModalities-compute_dice_distances"></a>}}
\if{latex}{\out{\hypertarget{method-ClustModalities-compute_dice_distances}{}}}
\subsection{Method \code{compute_dice_distances()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClustModalities$compute_dice_distances(disj_mat)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClustModalities-fit"></a>}}
\if{latex}{\out{\hypertarget{method-ClustModalities-fit}{}}}
\subsection{Method \code{fit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClustModalities$fit(X, k = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClustModalities-compute_elbow"></a>}}
\if{latex}{\out{\hypertarget{method-ClustModalities-compute_elbow}{}}}
\subsection{Method \code{compute_elbow()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClustModalities$compute_elbow(k_max = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClustModalities-plot_elbow"></a>}}
\if{latex}{\out{\hypertarget{method-ClustModalities-plot_elbow}{}}}
\subsection{Method \code{plot_elbow()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClustModalities$plot_elbow(k_max = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClustModalities-elbow"></a>}}
\if{latex}{\out{\hypertarget{method-ClustModalities-elbow}{}}}
\subsection{Method \code{elbow()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClustModalities$elbow(k_max = 10, plot = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{k_max}}{Maximum number of clusters to consider (default: 10)}

\item{\code{plot}}{Logical, whether to display the plot (default: TRUE)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
List with optimal_k, results data.frame, and plot function
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClustModalities-plot_dendrogram"></a>}}
\if{latex}{\out{\hypertarget{method-ClustModalities-plot_dendrogram}{}}}
\subsection{Method \code{plot_dendrogram()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClustModalities$plot_dendrogram(k = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClustModalities-plot_factor_map"></a>}}
\if{latex}{\out{\hypertarget{method-ClustModalities-plot_factor_map}{}}}
\subsection{Method \code{plot_factor_map()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClustModalities$plot_factor_map(dim1 = 1, dim2 = 2, show_labels = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClustModalities-plot_scree"></a>}}
\if{latex}{\out{\hypertarget{method-ClustModalities-plot_scree}{}}}
\subsection{Method \code{plot_scree()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClustModalities$plot_scree(cumulative = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClustModalities-plot_contrib"></a>}}
\if{latex}{\out{\hypertarget{method-ClustModalities-plot_contrib}{}}}
\subsection{Method \code{plot_contrib()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClustModalities$plot_contrib(dim = 1, top = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClustModalities-project_numeric"></a>}}
\if{latex}{\out{\hypertarget{method-ClustModalities-project_numeric}{}}}
\subsection{Method \code{project_numeric()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClustModalities$project_numeric(X_quant, plot = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClustModalities-cluster_table"></a>}}
\if{latex}{\out{\hypertarget{method-ClustModalities-cluster_table}{}}}
\subsection{Method \code{cluster_table()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClustModalities$cluster_table(k = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClustModalities-predict"></a>}}
\if{latex}{\out{\hypertarget{method-ClustModalities-predict}{}}}
\subsection{Method \code{predict()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClustModalities$predict(X_new)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClustModalities-print"></a>}}
\if{latex}{\out{\hypertarget{method-ClustModalities-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClustModalities$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClustModalities-summary"></a>}}
\if{latex}{\out{\hypertarget{method-ClustModalities-summary}{}}}
\subsection{Method \code{summary()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClustModalities$summary(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClustModalities-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ClustModalities-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClustModalities$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
